requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
cachetools==5.3.2
certifi==2023.11.17
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distlib==0.3.7
dnspython==2.4.2
email-validator==2.1.0
fastapi==0.104.1
filelock==3.13.1
google-api-core==2.14.0
google-auth==2.24.0
google-cloud-core==2.3.3
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
greenlet==3.0.1
h11==0.14.0
httptools==0.6.1
identify==2.5.32
idna==3.4
nodeenv==1.8.0
passlib==1.7.4
platformdirs==4.0.0
pre-commit==3.5.0
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
protobuf==4.25.1
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
PyJWT==2.8.0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
ruff==0.1.6
sentry-sdk==1.39.1
setuptools==69.0.2
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
uvloop==0.19.0
virtualenv==20.24.7
watchfiles==0.21.0
websockets==12.0
